chore: rename text_view.dart to text_shimmer.dart
